finish executing all traces of notebooks
The notebooks module/tests is removed and occurrences are changed to use
volumes. Additionally, my database was upgraded to new ID numbers and
the schema was changed to remove the notebooks table and use 'vid' in
place of 'nid' in the occs table. A log of the nid->vid translation has
been added to commit in case anything strange happens with the
references in the future (a quick check seems to indicate I didn't do
anything majorly and totally wrong, and I scanned the translation table
and my queries several times to be sure, but I'd rather have the log of
what I did just in case).

Next steps:
- Handle deleting all sorts of objects correctly -- I think we now have
  the necessary framework in place to write a couple of functions and
  take care of it.

- Implement the adding of occurrences so that entries+occurrences can be
  added. Also, make the "add" button in the search bar work & function
  correctly.

- Add printing support? It's not terribly urgent, but that would be the
  one remaining task that you can do in the old program that you can't
  in the new.

- Create the high-power multi-add window: that'll be a chore.