implement UOF (Unified Occurrence Format)
Good text parsers are the salt of the earth, but they sure as heck
ain't fun to implement. Or at least easy; it wasn't a bad programming
challenge, as they go. No segfaults involved, for instance, just a lot
of experimentation and a smattering of algorithm design.

It's also a good reminder that computers really, really suck at
understanding something that would be easily understandable to a human:
Unified Occurrence Format is not really very complicated, but the
parser is still some 200 lines of code before it interprets everything
in the format correctly.

We need to modify the database a bit to get the occurrences fields to
use slightly modified formats in some places (ranges and redirects).
Then we need to write a function that can use the UOF parser to create
an occurrence in combination with the UI, and then we'll be in business
as far as the backend for adding occurrences goes.