add entry deletion; event filter for shortcut enablement
On the second topic, the aboutToShow() enable trick only works well if
you don't care about keyboard shortcuts being enabled and disabled with
it, as that method is not called when running a keyboard shortcut. One
way would be to manually change all the enabled states whenever
something happened that changed them, but since there was no noticeable
performance degradation when I did so, I just went ahead and installed
an event filter that checks all of them whenever a modifier key is
pressed. This will certainly work fine for now, and will almost
certainly be fine in perpetuity.

Also begin working on implementing some more stuff in the occurrence
menu; since we're using a listWidget rather than a view, the next thing
to work on is changing the occurrence format to something stable and
pretty so that we can fetch an occurrence from the text displayed and do
the right thing with it.