add statusbar
Or, more correctly, "notice that you had a statusbar the whole time and
can now add something to it." It was basically invisible in the designer
and the implemented window!

Add statustips for the implemented menu items and a few others, and add
a permanent right-side widget that counts the entries and occurrences
displayed in the current view.