fix perceived parser bug (not actually one)
Added several tests for things that *should* actually work, and
clarified the documentation. One of the basic limits of the parser is
that it's not happy if there are numbers in the reference and you leave
out the colon or braces, and the only reason it seemed like this one
should be valid is that there wasn't in fact a number since it was a
single-volume one. There's not strictly any reason why it *couldn't* be
made to parse this situation as well, but there's no reason it
absolutely needs to, and frankly the code is enough of a mess already
without trying to parse sillily complex input.