implement password-protect feature
This doesn't provide any *real* security as in encrypting the database,
but it does lock out casual snoopers, who are not going to go poking
through your sqlite database to figure out how to disable the password
out of a pickled dictionary.

One possible security enhancement would be to make the "no password"
condition some goofy hard-coded hashed string; if you attempted to
manually set the password in the db to a blank string, it could still
refuse to start. (Or, for the hard-core people, it could self-destruct!
Not that I would ever want to use such a feature, and I'm not being
sarcastic here.)