implement entry classification tool
It's my first tool! And it works nicely. Just 5300 entries left to
classify :-).

A possible improvement would be user-specified categories, but that
would make the classification interface considerably harder to handle,
as it could no longer cleanly be implemented with radio buttons (they're
tough to add programmatically, plus what if someone added 50
categories?). If we don't do that, an "event" classification might be a
nice one to add (but we don't want to have 7 if we can avoid it, that's
a nasty number to try to line up anywhere).