begin simplification view; improve occurrence sorting
Occurrence sorting was ugly before; it relied on using some custom
functions and had been largely hacked together from a somewhat
sloppy-looking recipe on the Internet. Further, it had a subtle but very
significant bug: if there were two occurrences that had the same string
representation, they would be conflated in the returned sorted list,
because a dictionary was used to keep track of the occurrences while the
sort ran. The new method uses the __cmp__ method of the actual
occurrences and makes sorting considerably more convenient as well as
less buggy.