implement range printing
The simplification view is finished up for now, partly because I'm tired
of writing it and partly because I have a lot of homework to do now! It
*does* now do what I want with ranges, however.

I've decided to ignore redirects under the theory that not only is there
no good place to put them, they're not really very helpful: their
purpose is to help you find what you're looking for if you thought of it
under a different name, but in the simplification you're *seeing* the
entry names already, so that shouldn't really be an issue.