print simplification correctly when ranges involved
A very subtle bug: I neglected to use modifiedRangeKey(occ) instead of
plain str(occ) in one and only one part of the occurrence-enumerating
code, which meant that a range occurrence would end up overwriting
non-range occurrences that came before it. If the range happened to have
the highest oid, everything would be fine; otherwise some or all of the
other occurrences for that location with lower oids would be lost.

Not something that's easy to find until beta testing, when you look for
a particular entry's location in the simplification to do a nearby
lookup and discover the entry isn't there!