make forward/back work with incremental search
This does make the code more complicated, but I don't think it's
particularly inelegant. The one ugly thing is that with non-incremental
search, searches are saved in onSearch(), whereas with incremental
search they are saved by a focus-lost handler.

The criterion of search-box focus lost also isn't absolutely perfect for
determining whether a search should be saved in the history, but I think
it does a tolerably good job (and it's hard to think of something
better).