edit occurrence volume and page number
Unlike the cheap stopgap "change page" option I added a little while
back, this allows you to switch between reftypes, as well as to change
the volume number.

There's no reason why we couldn't make it change source as well, but I
think that's a very unlikely mistake, so having that option is not
high-priority at the moment.