bugfix: can't go back on following a redirect as first action
Additionally, it only happened so in incremental mode.

This was a very obscure bug that proved quite difficult to fix; the
solution was to make it permissible to go back even with an empty stack
if the value in the search box is non-empty.

Surprisingly, I didn't have to change anything in onGoBack to make this
work.