add new/open database options
Also wrote a little dialog to be displayed on first run or if the
database is moved, or (in the future) if the database isn't open for
some other odd reason. It has an option to open the last-used database,
but as of now I don't think it's possible to reach it since otherwise
that would just be opened.

I don't think there's reason to include a "close database" option since
the only options you would get upon doing that would be to open one or
create a new one, which you can already do perfectly fine from the file
menu, but that would be an option if we want to have it be one.