finish port to python3
This involved working with the ui side and making a couple of tweaks to
the db. I think the interface entirely works, and it's unlikely that any
subtle bugs will have been introduced by the 2->3 conversion.

The big changes:

- I got to get rid of the "unicode(whatever.text())" bullcrap; the new
API is just whatever.text() and is automatically converted to a Unicode
string.

- Similarly, no QVariant() or .toString() anymore, you just return the
data type or None, and things are converted to the correct type for you.

The subtle changes:

- No more __cmp__() (it's just silently ignored, it seems like the
interpreter could at least throw a warning on stderr), so occurrence
sort broke; I replaced it with __lt__(), which was not difficult.

- NoneType is not comparable and apparently it was before, so I had to
move the handling for cases where there are no occurrences closer to the
start in the database functions to fetch the first/last edited/added
dates. This is cleaner anyway.

- designer/makeForms.sh used implicit relative imports; it now uses
explicit ones. Same with the tests for the tests.utils module.

- The ./tabularium start file now explicitly calls python3 in the
shebang.

Overall this was a really painless port, and it's going to be nice not
to have to deal with the PITA that is Unicode handling in Python 2.