upgrade to Qt 5
Aside from a small tweak to QFileDialog's return value, the only thing I
had to do was change a bunch of imports from QtGui to QtWidgets.

Also put in a fix for the weird string error I was sometimes getting
when reading from the conf table; all I did was add TypeError to the
list of exceptions that mean there's no conf and it should init a new
one.

For some reason pylint is wacking out on me after I installed the pip3
version ("invalid method name __init__", for example). I'll take a look
at fixing that soon.