improve keyboard shortcuts
I'm not entirely happy with everything, but this is better than it was
before. Particular beefs:

- I wish there were a way not to redefine things like ^A and ^N. But
"new database" barely *ever* gets used and "notes" gets used much more,
so it seems stupid to keep that. ^A has the problem that you have to tab
out of the search box before you can use it, which is unfortunate, but
can I really use something different for "add"?

- "Jump" -> ^J is a nice shortcut, but the name is ugly where it gets
used.

- Alt-S for search, ^F for find. I might change "search" to "find" in
the interface to get rid of this problem; but there will still be an
Alt-F/Ctrl-F problem where Ctrl should be used because find is a common
shortcut in people's fingers but then it doesn't match the rest of the
Go menu.

I plan to assign function keys to the items in the tools menu, though
this does have the problem that if I put them alphabetically it will
look dumb if I add a tool later, or I'll have to break all users' muscle
memory. I may organize them into categories, if there can be some for a
short list. Or I could use no shortcuts except menu access keys for
those -- they're auxiliary tools after all.