fix selection of source limits
When making it so that it would be updated properly if the list of
sources was edited, I inadvertently made it impossible to select a
source, because the signal of selecting something was emitted and it
triggered the contents of the combo box to be reset.

We now reset the contents only when closing the source manager, and we
save the user's selection (or return to 'all' if the source is deleted).