properly sort entries and occs in tpeek
Also a few other tweaks to make find more useful.

I noticed that tpeek bypasses the password check: we may wish to add a
check and also an option to *permit* bypassing in tpeek in the
preferences dialog. In all, this really isn't a big deal -- if someone
knows enough to use tpeek, they could probably bypass the password check
without too much trouble anyway, not to mention that it's harder to
browse in a snoopy manner with the command-line functions.