fix tests, don't use INNER JOIN in find unless necessary
The tests had broken (I really need to set up a pre-commit hook here)
because I added occurrence filters, and this means that orphaned entries
will never be returned. The tests use a good share of orphaned entries
because it's a big waste of time to add occurrences when they're
completely irrelevant to the test.

We now use a different query if all of the occurrence filters passed to
find() are NONE.