completely rewrite large swaths of UOF handling
- The format has changed slightly; there are no longer braces, and
multiple pages of the same source are separated with ';', not ','. This
makes things easier to type, greatly reduces the chance of needing to
use an escape (when do you use a semicolon in an entry name?), and makes
things marginally easier to parse. It does increase typing somewhat if
you're adding multiple volumes of a single source, but *I've* never used
that and I highly doubt it's very generally useful.

- The parser now loops through all source names trying to find a source
at the beginning of UOF with .startswith(), rather than bending over
backwards with regex matches that often don't work correctly to figure
out if it's valid before even checking to see if the source exists. This
theoretically has higher asymptotic complexity (since we now have to do
an O(n) search over all our sources), but the previous method, besides
being prone to errors, quite possibly had a high enough constant-time
portion to take *longer* with a typical number of sources.

- There are now actually helpful error messages if the parser can't
figure out what you typed -- since the parser now knows itself what went
wrong, as it didn't before!