properly restore regex/incremental state on reload
These have signals that do the config change, and we blocked the signals
in the interest of speed.

I'm just slapping a Band-Aid on this particular issue for now, and I may
well be missing similar issues and have not tried to find them; a real
audit of the startup process and signal interdependencies is long
overdue if I want to increase general performance and prevent these
kinds of bugs. There's got to be a way to refactor all this.