automatically wash sort key entries
The "copy" and "wash" buttons are replaced by a single togglable "auto"
button, which starts out on and is turned off by clicking on it or by
changing the text of the sort key box manually. The signal that copies
the contents of the name box into the sort key box now also runs the
washing function.

Toggling "auto" back on serves the same function as the former "wash"
button. The one-button system is slightly less functional, but a lot
more intuitive, and the automatic washing makes up for a lot.

Sadly, the term "wash" is no longer user-facing. But that's probably a
good thing for user-friendliness, even if a bit sad!