add "integrations" for rtgrep and rpfind
Eventually my intention is to have an interface in the preferences
whereby the user can add their own integrations (you type in a command
line to be run and can substitute in the selected entry or occurrence or
nearby or search terms, say, with a printf sort of format string). Right
now I've just hard-coded these examples (and useful ones!).