add names-to-faces dialog
I'm partway through the long process of adding these for everyone in my
database now. I want to get it done at least for the STO folks before I
lose access to Stalkernet, so getting started on it now.

It should be noted that this update can markedly increase the size of
the database, since it will contain many embedded image files. That
said, the kind of images you usually download from the web to put in
here aren't usually all that large, so it's not going to become
gigabytes large or anything like that.

Could use some incremental improvements, like:

- in the manner of what I want for the Integrations menu, make the
image search functions user-configurable
- filter box at the top
- allow entries to be "hidden" that should not normally get a picture?
- check box to show only those entries that actually have pictures