implement cache for entries list
Surprisingly, the database hit for selecting an entry doesn't appear to
actually be the problem, though. I was pretty sure it was. Guess I
didn't analyze the problem that well...need to take a closer look at
what's firing off I guess, as if there's no database hit for the entry
literally the only thing that has to be done is repopulate the
occurrences/nearby/inspect areas, and that's taking like 1/4 second --
maybe the nearby/inspect database hits are really poor performing?

We may or may not want to toss this commit before merging into master,
seeing as it didn't seem to help much of anything. Probably better to
it than toss it though, as there are some good refactors in here too.