implement basic fts5 searching
This is significantly faster than substring searching and seems likely
to be nearly as effective. I might consider adding a raw substring mode
back in for the rare times when it might make a difference (although
simply falling back to regex for that would be reasonable and only
very slightly slower).