fix sort key being incorrectly set to manual mode
When doing a ^B or adding from the top bar, automatic mode was being
wrongly disabled in some cases because an explicit sort key was provided.
Instead we want to provide nothing, and change the condition for
entering manual mode so that it can only trigger if we knew a
sort key when setting up the dialog -- if the user has never provided a
sort key for this entry, it should *always* start in automatic mode.